api and  clean code
