+++ lib/src/api/provinders/data_provinder.dart
@@ -1,0 +1,8 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+import '../responses/service/data_service.dart';
+
+final dataProvinder = FutureProvider.autoDispose((_) {
+  final dataService = DataService.instance;
+  return dataService.getData();
+});
