+++ lib/src/api/responses/models/base_crypto_response.dart
@@ -1,6 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
 
-part 'base_crypto_response.g.dart';
+part '../base_crypto_response.g.dart';
 
 @JsonSerializable()
 class BaseDataModel {
