+++ lib/src/api/responses/models/data_crypto_response.dart
@@ -1,8 +1,8 @@
-import 'package:app1/api/models/base_crypto_response.dart';
-
 import 'package:json_annotation/json_annotation.dart';
 
-part 'data_crypto_response.g.dart';
+import '../models/base_crypto_response.dart';
+
+part '../data_crypto_response.g.dart';
 
 @JsonSerializable()
 class DataModel {
